# Optimize calendar date conversion