@statementreply 2021-02-11
[[time.cal.ymd.members]/19](https://eel.is/c++draft/time.cal.ymd.members#19):
> _Returns_: If `ok()`, [...]. Otherwise, if `y_­.ok() && m_­.ok()` is `true`, returns `sys_­days{y_­/m_­/1d} + (d_­ - 1d)`. Otherwise [...]

- Changed the type of `_Day_of_year` from `unsigned int` to `int` to clarify that it could be negative.
- Fixed comments on the range of intermediate variables.
- Added test cases for this clause.