@StephanTLavavej 2021-02-12
Thanks for significantly optimizing these core algorithms so `<chrono>` will consume less time! :stopwatch: :smiley_cat: :rocket: